Handle dynamic back urls
I have a url that was fetched from collection data. I was rendering the template before the subscription was done and the template data was {url: null}. 

Quick fix was to only render the template when the url existed but that cased a flicker when going though routes quickly.

This fix could have been done in an autorun but given that it is only called once this makes the most sense to me.