Update ionTabs.js
According to Ionic documentation:
"For iOS, tabs will appear at the bottom of the screen. For Android, tabs will be at the top of the screen, below the nav-bar."
http://ionicframework.com/docs/api/directive/ionTabs/

This change makes meteoric more consistent with Ionic making the UX closer to the native one.
This behavior can be prevented by passing any string to the style attribute, e.g. style="none", like in Ionic $ionicConfigProvider
http://ionicframework.com/docs/api/provider/$ionicConfigProvider/#tabs.style