Only scroll entering nav view, not the one you're leaving.
If you had already been to a page previously, and then navigated away from it, there was a flash of it's previous scroll position that showed.  This ensures that only the view that is being animated in to view, is scrolled to the correct position, and the one leaving the view is left alone.